선호 배우 및 전체 경로 수정

+++ video_test2.py
@@ -1,16 +1,6 @@
-# 영화 영상 분석 코드 - 사물인식으로 장르 판별, 유명인사 인식으로 배우 판별, 장면에서 배우 감정 판별
-# python, aws
-import csv
-
 import boto3, json, sys, time
 
 class VideoDetect:
-    with open('/Users/bag-yejin/Documents/credentials.py', 'r') as input:
-        next(input)
-        reader = csv.reader(input)
-        for line in reader:
-            access_key_id = line[0]
-            secret_access_key = line[1]
     jobId = 'null' # 비디오 분석 작업용 ID, 작업 식별자
     access_key_id = ''
     secret_access_key = ''
@@ -18,12 +8,10 @@ class VideoDetect:
     sqs = boto3.client('sqs', region_name='ap-northeast-2', aws_access_key_id=access_key_id, aws_secret_access_key=secret_access_key)
     sns = boto3.client('sns', region_name='ap-northeast-2', aws_access_key_id=access_key_id, aws_secret_access_key=secret_access_key)
 
-    roleArn = 'arn:aws:iam::392553513869:role/serviceRekognition'
-    bucket = 'allonsybuckets'
-    video = 'avengers.mp4'
-    startJobId1 = 'null'
-    startJobId2 = 'null'
-    startJobId3 = 'null'
+    roleArn =  'arn:aws:iam::392553513869:role/serviceRekognition'
+    bucket = 'allonsybucket1'
+    video = 'emotion_test.mp4'
+    startJobId = 'null'
 
     sqsQueueUrl = 'https://sqs.ap-northeast-2.amazonaws.com/805057381367/allonsySQS'
     snsTopicArn = 'arn:aws:sns:ap-northeast-2:805057381367:allonsySNS'
@@ -43,11 +31,11 @@ class VideoDetect:
             if sqsResponse:
                 if 'Messages' not in sqsResponse:
                     if dotLine < 40:
-                        #print('.', end='')
+                        print('.', end='')
                         dotLine = dotLine + 1
                         time.sleep(0.5)
                     else:
-                        #print()
+                        print()
                         dotLine = 0
                     sys.stdout.flush()
                     continue
@@ -55,12 +43,13 @@ class VideoDetect:
                 for message in sqsResponse['Messages']:
                     notification = json.loads(message['Body'])
                     rekMessage = json.loads(notification['Message'])
-                    jobFound = True
-                    if (rekMessage['Status'] == 'SUCCEEDED'):
-                        succeeded = True
-                    self.sqs.delete_message(QueueUrl=self.sqsQueueUrl, ReceiptHandle=message['ReceiptHandle'])
-                    # Delete the unknown message. Consider sending to dead letter queue
-                    self.sqs.delete_message(QueueUrl=self.sqsQueueUrl, ReceiptHandle=message['ReceiptHandle'])
+                    if rekMessage['JobId'] == self.startJobId:
+                        jobFound = True
+                        if (rekMessage['Status'] == 'SUCCEEDED'):
+                            succeeded = True
+                        self.sqs.delete_message(QueueUrl=self.sqsQueueUrl, ReceiptHandle=message['ReceiptHandle'])
+                        # Delete the unknown message. Consider sending to dead letter queue
+                        self.sqs.delete_message(QueueUrl=self.sqsQueueUrl, ReceiptHandle=message['ReceiptHandle'])
         return succeeded
 
     def GetLabelDetectionResults(self, second):
@@ -73,8 +62,9 @@ class VideoDetect:
         adventure = ['Train','Vehicle','Transportation','Nature']
         fscount = 0 # 3 이상되면 sf 장르로 분류
         adventurecount = 0
+
         while finished == False:
-            response = self.rek.get_label_detection(JobId=self.startJobId1, MaxResults=maxResults,
+            response = self.rek.get_label_detection(JobId=self.startJobId, MaxResults=maxResults,
                                                     NextToken=paginationToken, SortBy='TIMESTAMP') # 시간순으로 정렬, SortBy TIMESTAMP
 
             for labelDetection in response['Labels']:
@@ -101,23 +91,34 @@ class VideoDetect:
             genreList.append('adventure')
         print(genreList)
 
-    def GetFaceDetectionResults(self,second):
+    def GetFaceDetectionResults(self,second): # 장면분석, 감정은 가장 크게 느낀 2가지만 가져오기.
             second = int(second / 1000)
-            emotionList = []
+            emotionList = ""
+            correctSchema = "" # 수행모델을 위한 10초 간격의 감정정보
+            timeList = []
             maxResults = 10
             paginationToken = ''
             finished = False
+            t=0
 
             while finished == False:
-                response = self.rek.get_face_detection(JobId=self.startJobId2, MaxResults=maxResults, NextToken=paginationToken)
+                response = self.rek.get_face_detection(JobId=self.startJobId, MaxResults=maxResults, NextToken=paginationToken)
 
                 for faceDetection in response['Faces']:
+                    time = faceDetection['Timestamp']//1000 # 초로 분류됨
+                    if time//10==t:
+                        t+=1
+                        timeList.append(time)
+                        print(faceDetection['Face']['Emotions'][0])
+                        print(faceDetection['Timestamp'])
+                        correctSchema += ((faceDetection['Face']['Emotions'][0])["Type"]) + " "
+
                     if faceDetection['Timestamp'] / 1000 >= 5:
                         last1 = second - 5
-
                         if int(faceDetection['Timestamp'] / 1000) <= second + 5 and int(faceDetection['Timestamp'] / 1000) >= last1:
-                            if str((faceDetection['Face']['Emotions'][0])["Type"]) not in emotionList:
-                                emotionList.append(str((faceDetection['Face']['Emotions'][0])["Type"]))
+                            if ((faceDetection['Face']['Emotions'][0])["Type"]) not in emotionList:
+                                if int((faceDetection['Face']['Emotions'][0])["Confidence"])>=80 and ((faceDetection['Face']['Emotions'][0])["Type"])!="CALM":
+                                    emotionList += (((faceDetection['Face']['Emotions'][0])["Type"]))+ " "
                         else:
                             continue
 
@@ -126,6 +127,7 @@ class VideoDetect:
                 else:
                     finished = True
             print(emotionList)
+            print(correctSchema)
 
     def CreateTopicandQueue(self):
 
@@ -181,25 +183,27 @@ class VideoDetect:
         self.sqs.delete_queue(QueueUrl=self.sqsQueueUrl)
         self.sns.delete_topic(TopicArn=self.snsTopicArn)
 
-    def StartDetection(self):
+    def StartDetection1(self):
         # 사물 인식 아이디 발급
         response1 = self.rek.start_label_detection(Video={'S3Object': {'Bucket': self.bucket, 'Name': self.video}},
                                                   NotificationChannel={'RoleArn': self.roleArn,
                                                                        'SNSTopicArn': self.snsTopicArn})
-        self.startJobId1 = response1['JobId']
+        self.startJobId = response1['JobId']
 
+    def StartDetection3(self):
         # 감정 인식 아이디 발급
         response2 = self.rek.start_face_detection(Video={'S3Object': {'Bucket': self.bucket, 'Name': self.video}},
                                                  NotificationChannel={'RoleArn': self.roleArn,
                                                                       'SNSTopicArn': self.snsTopicArn},
                                                  FaceAttributes='ALL')
-        self.startJobId2 = response2['JobId']
+        self.startJobId = response2['JobId']
 
+    def StartDetection2(self):
         # 유명인 인식 아이디 발급
         response3 = self.rek.start_celebrity_recognition(
             Video={'S3Object': {'Bucket': self.bucket, 'Name': self.video}},
             NotificationChannel={'RoleArn': self.roleArn, 'SNSTopicArn': self.snsTopicArn})
-        self.startJobId3 = response3['JobId']
+        self.startJobId = response3['JobId']
 
     def GetCelebrityDetectionResults(self,second):
         second = int(second / 1000)
@@ -208,7 +212,7 @@ class VideoDetect:
         finished = False
         celeblist = []
         while finished == False:
-            response = self.rek.get_celebrity_recognition(JobId=self.startJobId3, MaxResults=maxResults,NextToken=paginationToken)
+            response = self.rek.get_celebrity_recognition(JobId=self.startJobId, MaxResults=maxResults,NextToken=paginationToken)
 
             for celebrityRecognition in response['Celebrities']:
                 if celebrityRecognition['Timestamp']/1000>=2:
@@ -227,16 +231,20 @@ class VideoDetect:
 
 def main():
     roleArn = 'arn:aws:iam::392553513869:role/serviceRekognition'
-    bucket = 'allonsybuckets'
-    video = 'avengers.mp4'
+    bucket = 'allonsybucket1'
+    video = 'emotion_test.mp4'
 
     analyzer = VideoDetect(roleArn, bucket, video)
     analyzer.CreateTopicandQueue()
 
-    analyzer.StartDetection()
+    #analyzer.StartDetection1()
+    #if analyzer.GetSQSMessageSuccess() == True:
+     #   analyzer.GetLabelDetectionResults(25000)  # 25초에 사용자 감정의 폭 Max
+    #analyzer.StartDetection2()
+    #if analyzer.GetSQSMessageSuccess() == True:
+     #   analyzer.GetCelebrityDetectionResults(72333)  # 72초에 사용자 감정의 폭 Max
+    analyzer.StartDetection3()
     if analyzer.GetSQSMessageSuccess() == True:
-        analyzer.GetLabelDetectionResults(25000)  # 25초에 사용자 감정의 폭 Max
-        analyzer.GetCelebrityDetectionResults(72333)  # 72초에 사용자 감정의 폭 Max
         analyzer.GetFaceDetectionResults(72333) # 72초에 사용자 감정의 폭 Max
 
     analyzer.DeleteTopicandQueue()
